+++ modules/characters_encoding/builtin/cpp/nativecharsetBuiltin.cpp
@@ -54,7 +54,7 @@ Nelson::CharactersEncodingGateway::nativecharsetBuiltin(int nLhs, const ArrayOfV
     if (isSupportedAsNumeric || isSupportedAsCharacterArray || isSupportedAsScalarString) {
         if (isSupportedAsNumeric) {
             param1.promoteType(NLS_UINT8);
-            indexType l = param1.getLength();
+            indexType l = param1.getElementCount();
             if (l > 0) {
                 uint8* s = (uint8*)param1.getDataPointer();
                 const char* chars = reinterpret_cast<const char*>(s);