+++ modules/string/builtin/cpp/stringsBuiltin.cpp
@@ -65,9 +65,9 @@ Nelson::StringGateway::stringsBuiltin(Evaluator* eval, int nLhs, const ArrayOfVe
                     retval.push_back(c);
                 } else {
                     ArrayOf arg = argIn[0];
-                    Dimensions dims(arg.getLength());
+                    Dimensions dims(arg.getElementCount());
                     auto* dindex = (double*)arg.getDataPointer();
-                    for (indexType k = 0; k < arg.getLength(); k++) {
+                    for (indexType k = 0; k < arg.getElementCount(); k++) {
                         double _dIndex = dindex[k];
                         if (!std::isfinite(_dIndex)) {
                             Error(ERROR_WRONG_ARGUMENT_1_FINITE_VECTOR_INTEGER_VALUE_EXPECTED);